Fixing gradle build issue
